+++ src/Mpociot/BotMan/Drivers/FacebookDriver.php
@@ -2,6 +2,7 @@
 
 namespace Mpociot\BotMan\Drivers;
 
+use Mpociot\BotMan\Interfaces\DriverInterface;
 use Mpociot\BotMan\User;
 use Mpociot\BotMan\Answer;
 use Mpociot\BotMan\Message;
@@ -166,7 +167,7 @@ class FacebookDriver extends Driver
      * @param array $additionalParameters
      * @return \Symfony\Component\HttpFoundation\Response
      */
-    public function reply($message, $matchingMessage, $additionalParameters = [])
+    public function buildServicePayload($message, $matchingMessage, $additionalParameters = [])
     {
         $parameters = array_merge_recursive([
             'recipient' => [
@@ -208,7 +209,16 @@ class FacebookDriver extends Driver
 
         $parameters['access_token'] = $this->config->get('facebook_token');
 
-        return $this->http->post('https://graph.facebook.com/v2.6/me/messages', [], $parameters);
+        return $parameters;
+    }
+
+    /**
+     * @param mixed $payload
+     * @return Response
+     */
+    public function sendPayload($payload)
+    {
+        return $this->http->post('https://graph.facebook.com/v2.6/me/messages', [], $payload);
     }
 
     /**