+++ src/Mpociot/BotMan/Drivers/NullDriver.php
@@ -71,8 +71,18 @@ class NullDriver extends Driver
      * @param array $additionalParameters
      * @return $this
      */
-    public function reply($message, $matchingMessage, $additionalParameters = [])
+    public function buildServicePayload($message, $matchingMessage, $additionalParameters = [])
     {
+        return null;
+    }
+
+    /**
+     * @param mixed $payload
+     * @return Response
+     */
+    public function sendPayload($payload)
+    {
+        return null;
     }
 
     /**