+++ src/Mpociot/BotMan/Drivers/SlackDriver.php
@@ -2,6 +2,8 @@
 
 namespace Mpociot\BotMan\Drivers;
 
+use Mpociot\BotMan\BotMan;
+use Mpociot\BotMan\Interfaces\DriverInterface;
 use Mpociot\BotMan\User;
 use Mpociot\BotMan\Answer;
 use Mpociot\BotMan\Message;
@@ -15,6 +17,12 @@ use Mpociot\BotMan\Messages\Message as IncomingMessage;
 class SlackDriver extends Driver
 {
     const DRIVER_NAME = 'Slack';
+
+    const RESULT_TOKEN = 'token';
+
+    const RESULT_JSON = 'json';
+
+    protected $resultType = self::RESULT_JSON;
 
     /**
      * @param Request $request
@@ -125,37 +133,55 @@ class SlackDriver extends Driver
      * @param string|Question $message
      * @param Message $matchingMessage
      * @param array $additionalParameters
-     * @return $this|void
+     * @return array
      */
-    public function reply($message, $matchingMessage, $additionalParameters = [])
+    public function buildServicePayload($message, $matchingMessage, $additionalParameters = [])
     {
         if (! Collection::make($matchingMessage->getPayload())->has('team_domain')) {
-            $this->replyWithToken($message, $matchingMessage, $additionalParameters);
+            $this->resultType = self::RESULT_TOKEN;
+            $payload = $this->replyWithToken($message, $matchingMessage, $additionalParameters);
         } else {
-            $this->respondJSON($message, $matchingMessage, $additionalParameters);
+            $this->resultType = self::RESULT_JSON;
+            $payload = $this->respondJSON($message, $matchingMessage, $additionalParameters);
         }
+        return $payload;
+    }
+
+    /**
+     * @param mixed $payload
+     * @return Response
+     */
+    public function sendPayload($payload)
+    {
+        if ($this->resultType == self::RESULT_TOKEN) {
+            return $this->http->post('https://slack.com/api/chat.postMessage', [], $payload);
+        }
+        return Response::create(json_encode($payload), 200, ['Content-Type', 'application/json'])->send();
+
     }
 
     /**
      * @param $message
      * @param array $additionalParameters
      * @param Message $matchingMessage
-     * @return $this
+     * @return array
      */
-    public function replyInThread($message, $additionalParameters, $matchingMessage)
+    public function replyInThread($message, $additionalParameters, $matchingMessage, BotMan $bot)
     {
         $additionalParameters['thread_ts'] = ! empty($matchingMessage->getPayload()->get('thread_ts'))
             ? $matchingMessage->getPayload()->get('thread_ts')
             : $matchingMessage->getPayload()->get('ts');
 
-        return $this->reply($message, $matchingMessage, $additionalParameters);
+        $payload = $this->buildServicePayload($message, $matchingMessage, $additionalParameters);
+
+        return $bot->sendPayload($payload);
     }
 
     /**
      * @param string|Question $message
      * @param Message $matchingMessage
      * @param array $parameters
-     * @return $this
+     * @return array
      */
     protected function respondJSON($message, $matchingMessage, $parameters = [])
     {
@@ -175,14 +201,14 @@ class SlackDriver extends Driver
             $parameters['text'] = $this->format($message);
         }
 
-        Response::create(json_encode($parameters), 200, ['Content-Type', 'application/json'])->send();
+        return $parameters;
     }
 
     /**
      * @param string|Question|IncomingMessage $message
      * @param Message $matchingMessage
      * @param array $additionalParameters
-     * @return Response
+     * @return array
      */
     protected function replyWithToken($message, $matchingMessage, $additionalParameters = [])
     {
@@ -208,7 +234,7 @@ class SlackDriver extends Driver
 
         $parameters['token'] = $this->config->get('slack_token');
 
-        return $this->http->post('https://slack.com/api/chat.postMessage', [], $parameters);
+        return $parameters;
     }
 
     /**