+++ src/Mpociot/BotMan/Drivers/TelegramDriver.php
@@ -16,6 +16,8 @@ class TelegramDriver extends Driver
 {
     const DRIVER_NAME = 'Telegram';
 
+    protected $endpoint = 'sendMessage';
+
     /**
      * @param Request $request
      */
@@ -156,9 +158,8 @@ class TelegramDriver extends Driver
      * @param array $additionalParameters
      * @return Response
      */
-    public function reply($message, $matchingMessage, $additionalParameters = [])
+    public function buildServicePayload($message, $matchingMessage, $additionalParameters = [])
     {
-        $endpoint = 'sendMessage';
         $parameters = array_merge_recursive([
             'chat_id' => $matchingMessage->getChannel(),
         ], $additionalParameters);
@@ -174,15 +175,15 @@ class TelegramDriver extends Driver
         } elseif ($message instanceof IncomingMessage) {
             if (! is_null($message->getImage())) {
                 if (strtolower(pathinfo($message->getImage(), PATHINFO_EXTENSION)) === 'gif') {
-                    $endpoint = 'sendDocument';
+                    $this->endpoint = 'sendDocument';
                     $parameters['document'] = $message->getImage();
                 } else {
-                    $endpoint = 'sendPhoto';
+                    $this->endpoint = 'sendPhoto';
                     $parameters['photo'] = $message->getImage();
                 }
                 $parameters['caption'] = $message->getMessage();
             } elseif (! is_null($message->getVideo())) {
-                $endpoint = 'sendVideo';
+                $this->endpoint = 'sendVideo';
                 $parameters['video'] = $message->getVideo();
                 $parameters['caption'] = $message->getMessage();
             } else {
@@ -192,7 +193,16 @@ class TelegramDriver extends Driver
             $parameters['text'] = $message;
         }
 
-        return $this->http->post('https://api.telegram.org/bot'.$this->config->get('telegram_token').'/'.$endpoint, [], $parameters);
+        return $parameters;
+    }
+
+    /**
+     * @param mixed $payload
+     * @return Response
+     */
+    public function sendPayload($payload)
+    {
+        return $this->http->post('https://api.telegram.org/bot'.$this->config->get('telegram_token').'/'.$this->endpoint, [], $payload);
     }
 
     /**