+++ tests/Drivers/BotFrameworkDriverTest.php
@@ -263,7 +263,8 @@ class BotFrameworkDriverTest extends PHPUnit_Framework_TestCase
         ], $html);
 
         $message = $driver->getMessages()[0];
-        $driver->reply('Test', $message);
+        $payload = $driver->buildServicePayload('Test', $message);
+        $driver->sendPayload($payload);
     }
 
     /** @test */
@@ -304,7 +305,8 @@ class BotFrameworkDriverTest extends PHPUnit_Framework_TestCase
 
         $user = '29:1zPNq1EP2_H-mik_1MQgKYp0nZu9tUljr2VEdTlGhEo7VlZ1YVDVSUZ0g70sk1';
         $message = new Message('hey there', $user, $user);
-        $driver->reply('Test', $message);
+        $payload = $driver->buildServicePayload('Test', $message);
+        $driver->sendPayload($payload);
     }
 
     /** @test */
@@ -345,9 +347,11 @@ class BotFrameworkDriverTest extends PHPUnit_Framework_TestCase
         ], $html);
 
         $message = $driver->getMessages()[0];
-        $driver->reply('Test', $message, [
+
+        $payload = $driver->buildServicePayload('Test', $message, [
             'foo' => 'bar',
         ]);
+        $driver->sendPayload($payload);
     }
 
     /** @test */
@@ -387,7 +391,8 @@ class BotFrameworkDriverTest extends PHPUnit_Framework_TestCase
         ], $html);
 
         $message = $driver->getMessages()[0];
-        $driver->reply(\Mpociot\BotMan\Messages\Message::create('Test'), $message);
+        $payload = $driver->buildServicePayload(\Mpociot\BotMan\Messages\Message::create('Test'), $message);
+        $driver->sendPayload($payload);
     }
 
     /** @test */
@@ -433,7 +438,8 @@ class BotFrameworkDriverTest extends PHPUnit_Framework_TestCase
         ], $html);
 
         $message = $driver->getMessages()[0];
-        $driver->reply(\Mpociot\BotMan\Messages\Message::create('Test')->image('http://foo.com/bar.png'), $message);
+        $payload = $driver->buildServicePayload(\Mpociot\BotMan\Messages\Message::create('Test')->image('http://foo.com/bar.png'), $message);
+        $driver->sendPayload($payload);
     }
 
     /** @test */
@@ -479,6 +485,7 @@ class BotFrameworkDriverTest extends PHPUnit_Framework_TestCase
         ], $html);
 
         $message = $driver->getMessages()[0];
-        $driver->reply(\Mpociot\BotMan\Messages\Message::create('Test')->video('http://foo.com/bar.mp4'), $message);
+        $payload = $driver->buildServicePayload(\Mpociot\BotMan\Messages\Message::create('Test')->video('http://foo.com/bar.mp4'), $message);
+        $driver->sendPayload($payload);
     }
 }