+++ tests/Drivers/SlackDriverTest.php
@@ -3,6 +3,7 @@
 namespace Mpociot\BotMan\Tests\Drivers;
 
 use Mockery as m;
+use Mpociot\BotMan\BotMan;
 use Mpociot\BotMan\Button;
 use Mpociot\BotMan\Message;
 use Mpociot\BotMan\Question;
@@ -14,6 +15,11 @@ use Symfony\Component\HttpFoundation\Request;
 
 class SlackDriverTest extends PHPUnit_Framework_TestCase
 {
+    public function tearDown()
+    {
+        m::close();
+    }
+
     private function getDriver($responseData, $htmlInterface = null)
     {
         $request = m::mock(Request::class.'[getContent]');
@@ -351,7 +357,7 @@ class SlackDriverTest extends PHPUnit_Framework_TestCase
         ], $html);
 
         $message = new Message('', '', 'general');
-        $driver->reply('Test', $message);
+        $driver->sendPayload($driver->buildServicePayload('Test', $message));
     }
 
     /** @test */
@@ -382,7 +388,7 @@ class SlackDriverTest extends PHPUnit_Framework_TestCase
         ], $html);
 
         $message = new Message('', '', 'general');
-        $driver->reply(\Mpociot\BotMan\Messages\Message::create('Test'), $message);
+        $driver->sendPayload($driver->buildServicePayload(\Mpociot\BotMan\Messages\Message::create('Test'), $message));
     }
 
     /** @test */
@@ -414,7 +420,7 @@ class SlackDriverTest extends PHPUnit_Framework_TestCase
         ], $html);
 
         $message = new Message('', '', 'general');
-        $driver->reply(\Mpociot\BotMan\Messages\Message::create('Test', 'http://image.url/foo.png'), $message);
+        $driver->sendPayload($driver->buildServicePayload(\Mpociot\BotMan\Messages\Message::create('Test', 'http://image.url/foo.png'), $message));
     }
 
     /** @test */
@@ -433,7 +439,7 @@ class SlackDriverTest extends PHPUnit_Framework_TestCase
             'text' => 'Hi Julia',
         ]);
         $driver = new SlackDriver($request, [], m::mock(Curl::class));
-        $driver->reply('test', $driver->getMessages()[0]);
+        $driver->sendPayload($driver->buildServicePayload('test', $driver->getMessages()[0]));
     }
 
     /** @test */
@@ -453,7 +459,7 @@ class SlackDriverTest extends PHPUnit_Framework_TestCase
             'text' => 'Hi Julia',
         ]);
         $driver = new SlackDriver($request, [], m::mock(Curl::class));
-        $driver->reply('test', $driver->getMessages()[0]);
+        $driver->sendPayload($driver->buildServicePayload('test', $driver->getMessages()[0]));
     }
 
     /** @test */
@@ -489,7 +495,7 @@ class SlackDriverTest extends PHPUnit_Framework_TestCase
         ], $html);
 
         $message = new Message('', '', 'general');
-        $driver->reply($question, $message);
+        $driver->sendPayload($driver->buildServicePayload($question, $message));
     }
 
     /** @test */
@@ -525,13 +531,13 @@ class SlackDriverTest extends PHPUnit_Framework_TestCase
         ], $html);
 
         $message = new Message('response', '', 'general');
-        $driver->reply('Test', $message, [
+        $driver->sendPayload($driver->buildServicePayload('Test', $message, [
             'username' => 'ReplyBot',
             'icon_emoji' => ':dash:',
             'attachments' => json_encode([[
                 'image_url' => 'imageurl',
             ]]),
-        ]);
+        ]));
     }
 
     /** @test */
@@ -571,13 +577,17 @@ class SlackDriverTest extends PHPUnit_Framework_TestCase
         $message = new Message('response', '', 'general', Collection::make([
             'ts' => '1234.5678',
         ]));
+
+        $botman = m::mock(BotMan::class)->makePartial();
+        $botman->setDriver($driver);
+
         $driver->replyInThread('Test', [
             'username' => 'ReplyBot',
             'icon_emoji' => ':dash:',
             'attachments' => json_encode([[
                 'image_url' => 'imageurl',
             ]]),
-        ], $message);
+        ], $message, $botman);
     }
 
     /** @test */